requirements.txt
langchain>=0.1.0
langchain-openai>=0.0.5
openai>=1.0.0
pytesseract>=0.3.10
Pillow>=10.0.0
pdf2image>=1.16.3
python-dotenv>=1.0.0
pydantic>=2.0.0
typing-extensions>=4.8.0

